backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
numpy>=1.24.0
faiss-cpu>=1.7.4
opencv-python>=4.8.0
Pillow>=10.0.0
librosa>=0.10.0
scikit-learn>=1.3.0
pydantic==2.10.3
supabase>=2.0.0
tensorflow>=2.15.0
keras-facenet>=0.3.2
soundfile>=0.12.1
speechbrain>=0.5.16
torch>=2.0.0
torchaudio>=2.0.0
scikit-image>=0.21.0
pydub>=0.25.1
